Files moved around and minor fixes
