+++ Umbraco.Frontend/.editorconfig
@@ -1,0 +1,6 @@
+[*]
+charset = utf-8
+end_of_line = crlf
+indent_size = 2
+indent_style = space
+insert_final_newline = true
