+++ Umbraco.Frontend/.eslintrc.js
@@ -10,7 +10,10 @@ module.exports = {
     'plugin:@typescript-eslint/recommended',
     "plugin:jsx-a11y/recommended",
   ],
-  "plugins": ["@typescript-eslint"],
+  "plugins": [
+    "react-hooks",
+    "@typescript-eslint",
+  ],
   parserOptions: {
     ecmaVersion: 2018,
     sourceType: 'module',  // Allows for the use of imports
@@ -21,6 +24,7 @@ module.exports = {
     },
   },
   rules: {
+    "@typescript-eslint/camelcase": "off",
     "@typescript-eslint/explicit-function-return-type": "off",
     "@typescript-eslint/explicit-member-accessibility": "off",
     "@typescript-eslint/indent": "off",
@@ -28,27 +32,49 @@ module.exports = {
     "@typescript-eslint/no-empty-interface": "off",
     "@typescript-eslint/no-explicit-any": "off",
     "@typescript-eslint/no-var-requires": "off",
+    "arrow-body-style": "off",
+    "arrow-parens": "off",
     "class-methods-use-this": "off",
     "comma-dangle": ["error", "always-multiline"],
+    "default-case": "off",
+    "dot-notation": "off",
+    "eol-last": "warn",
+    "import/newline-after-import": "warn",
     "import/no-unresolved": "off",
     "import/no-webpack-loader-syntax": "off",
     "import/prefer-default-export": "off",
-    "indent": ["error", 2],
+    "indent": ["warn", 2],
     "jsx-a11y/accessible-emoji": "off",
     "jsx-a11y/label-has-associated-control": "off",
+    "jsx-a11y/no-onchange": "off",
     "linebreak-style": ["error", "windows"],
     "lines-between-class-members": ["error", "always", { exceptAfterSingleLine: true }],
+    "max-len": "warn",
     "no-confusing-arrow": "off",
+    "no-multi-spaces": "warn",
+    "no-multiple-empty-lines": "warn",
+    "no-nested-ternary": "off",
+    "no-param-reassign": "off",
     "no-plusplus": "off",
+    "no-trailing-spaces": "warn",
+    "no-underscore-dangle": "off",
     "object-curly-newline": "off",
+    "object-curly-spacing": "warn",
     "padded-blocks": "off",
+    "quotes": ["warn", "double", { "avoidEscape": true, "allowTemplateLiterals": true }],
+    "react-hooks/exhaustive-deps": "warn",
+    "react-hooks/rules-of-hooks": "error",
     "react/button-has-type": "off",
     "react/destructuring-assignment": "off",
     "react/jsx-filename-extension": "off",
+    "react/jsx-max-props-per-line": "off",
     "react/jsx-one-expression-per-line": ["on", { "allow": "none" | "literal" | "single-child" }],
     "react/no-array-index-key": "off",
     "react/prefer-stateless-function": "off",
-    "react/prop-types": [2, { ignore: ['children'] }],
+    "react/prop-types": "off",
+    "react/sort-comp": "off",
     "semi": ["error", "always"],
+    "spaced-comment": "off",
+    "import/order": "off",
   },
 };
