+++ Umbraco.Frontend/package.json
@@ -23,7 +23,7 @@
     "eslint-plugin-react": "^7.12.4",
     "fast-async": "^6.3.8",
     "file-loader": "^3.0.1",
-    "mini-css-extract-plugin": "^0.5.0",
+    "mini-css-extract-plugin": "^0.6.0",
     "node-sass": "^4.11.0",
     "node-sass-glob-importer": "^5.3.1",
     "optimize-css-assets-webpack-plugin": "^5.0.1",
@@ -32,9 +32,9 @@
     "sass-loader": "^7.1.0",
     "sass-resources-loader": "^2.0.0",
     "style-loader": "^0.23.1",
-    "stylelint": "^9.10.1",
-    "stylelint-config-primer": "^6.0.0",
-    "stylelint-order": "^2.1.0",
+    "stylelint": "^10.0.1",
+    "stylelint-config-primer": "^7.0.0",
+    "stylelint-order": "^3.0.0",
     "terser-webpack-plugin": "^1.2.2",
     "typescript": "^3.3.4000",
     "url-loader": "^1.1.2",
@@ -46,7 +46,7 @@
   "scripts": {
     "build": "webpack --config ./webpack --env.production",
     "clean:mac": "rm -rf node_modules && npm i",
-		"clean:pc": "rd /s /q node_modules && npm i",
+    "clean:pc": "rd /s /q node_modules && npm i",
     "clean:cache": "rm -rf [package-lock.json] node_modules && npm cache clean -f && npm i",
     "dev": "webpack-dev-server --config ./webpack --inline --progress --hot",
     "start": "yarn dev"
