+++ Umbraco.Frontend/src/styles/layout/_header.scss
