+++ Umbraco.Frontend/src/styles/resources/functions/color.scss
