+++ Umbraco.Frontend/src/styles/resources/helpers/media-label.scss
