+++ Umbraco.Frontend/src/styles/resources/mixins/icon-font.scss
