+++ Umbraco.Frontend/src/styles/resources/variables/colors.scss
