+++ Umbraco.Frontend/src/types/window.d.ts
