+++ Umbraco.Frontend/webpack/optimization.js
