--- tsconfig.json
@@ -1,40 +1,0 @@
-{
-  "compilerOptions": {
-    "allowJs": true,
-    "jsx": "react",
-    "baseUrl": "./",
-    "target": "es5",
-    "module": "esnext",
-    "lib": [
-      "dom",
-      "es5",
-      "es2015.promise"
-    ],
-    "forceConsistentCasingInFileNames": false,
-    "importHelpers": true,
-    "moduleResolution": "node",
-    "noImplicitAny": false,
-    "noUnusedLocals": true,
-    "noImplicitReturns": true,
-    "noImplicitThis": true,
-    "preserveSymlinks": false,
-    "sourceMap": true,
-    "suppressImplicitAnyIndexErrors": true,
-    "strictNullChecks": false,
-    "outDir": "build",
-    "rootDir": "../",
-    "paths": {
-      "helpers/*": [
-        "src/utils/*"
-      ],
-      "components/*": [
-        "src/components/*"
-      ],
-      "interfaces/*": [
-        "src/interfaces/*"
-      ]
-    },
-  },
-  "exclude": ["node_modules", "build", "scripts", "webpack"],
-  "typeRoots": ["node_modules/@types"]
-}
